LayoutTests/inspector/console/console-format-style-expected.txt

CONSOLE MESSAGE: line 8: %cBlue!.
CONSOLE MESSAGE: line 9: %cRed!%c.
Tests that console logging dumps properly styled messages, and that the whole message gets the same style, regardless of multiple %c settings.

Blue!. console-format-style.html:8
Style: color: blue;
Red!. console-format-style.html:9
Style: color: red;
